Add PUT and GET methods for /api/products. Handle invalid ID's for PUT and DELETE.

+++ backend/server.js
@@ -2,12 +2,23 @@ import express from "express";
 import dotenv from "dotenv";
 import { connectDB } from "./config/db.js";
 import Product from "./models/product.model.js";
+import mongoose from "mongoose";
 
 dotenv.config();
 
 const app = express();
 
 app.use(express.json());
+
+app.get("/api/products", async (req, res) => {
+  try {
+    const products = await Product.find({});
+    res.status(200).json({ success: true, data: products });
+  } catch (error) {
+    console.log(`Error fetching products: ${error.message}`);
+    return res.status(500).json({ success: false, message: "Server error" });
+  }
+});
 
 app.post("/api/products", async (req, res) => {
   const product = req.body;
@@ -26,17 +37,45 @@ app.post("/api/products", async (req, res) => {
     res.status(201).json({ success: true, data: newProduct });
   } catch (error) {
     console.error(`Could not create product: ${error.message}`);
-    res.status(500).json({ success: false, message: "Server Error" });
+    res.status(500).json({ success: false, message: "Server error" });
+  }
+});
+
+app.put("/api/products/:id", async (req, res) => {
+  const { id } = req.params;
+  const product = req.body;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product ID" });
+  }
+
+  try {
+    const updatedProduct = await Product.findByIdAndUpdate(id, product, {
+      new: true,
+    });
+    res.status(200).json({ success: true, data: updatedProduct });
+  } catch (error) {
+    console.error(`Could not update product: ${error.message}`);
+    res.status(500).json({ success: false, message: "Server error" });
   }
 });
 
 app.delete("/api/products/:id", async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product deleted" });
   } catch (error) {
+    console.error(`Could not delete product: ${error.message}`);
     res.status(404).json({ success: false, message: "Product not found" });
   }
 });
